Move hex_props to elem_attr.hpp
